+++ Riot/Modules/ServiceTerms/Modal/Modal/ServiceTermsModalScreenViewModelType.swift
@@ -33,6 +33,9 @@ protocol ServiceTermsModalScreenViewModelType {
 
     var serviceUrl: String { get }
     var serviceType: MXServiceType { get }
+    /// If true, terms are displayed out of a context of a flow (like a background 3pids lookup)
+    /// In this case, the wording needs to provide more information about the intent
+    var outOfContext: Bool { get }
     var policies: [MXLoginPolicyData]? { get set }
     var alreadyAcceptedPoliciesUrls: [String] { get set }