+++ Riot/Modules/ServiceTerms/Modal/ServiceTermsModalCoordinatorBridgePresenter.swift
@@ -35,6 +35,7 @@ final class ServiceTermsModalCoordinatorBridgePresenter: NSObject {
     private let session: MXSession
     private let baseUrl: String
     private let serviceType: MXServiceType
+    private let outOfContext: Bool
     private let accessToken: String
     private var coordinator: ServiceTermsModalCoordinator?
     
@@ -48,10 +49,11 @@ final class ServiceTermsModalCoordinatorBridgePresenter: NSObject {
     
     // MARK: - Setup
     
-    init(session: MXSession, baseUrl: String, serviceType: MXServiceType, accessToken: String) {
+    init(session: MXSession, baseUrl: String, serviceType: MXServiceType, outOfContext: Bool = false, accessToken: String) {
         self.session = session
         self.baseUrl = baseUrl
         self.serviceType = serviceType
+        self.outOfContext = outOfContext
         self.accessToken = accessToken
         super.init()
     }
@@ -64,7 +66,7 @@ final class ServiceTermsModalCoordinatorBridgePresenter: NSObject {
     // }
     
     func present(from viewController: UIViewController, animated: Bool) {
-        let serviceTermsModalCoordinator = ServiceTermsModalCoordinator(session: self.session, baseUrl: self.baseUrl, serviceType: self.serviceType, accessToken: accessToken)
+        let serviceTermsModalCoordinator = ServiceTermsModalCoordinator(session: self.session, baseUrl: self.baseUrl, serviceType: self.serviceType, outOfContext: self.outOfContext, accessToken: accessToken)
         serviceTermsModalCoordinator.delegate = self
         viewController.present(serviceTermsModalCoordinator.toPresentable(), animated: animated, completion: nil)
         serviceTermsModalCoordinator.start()